+++ pkg/proxy/ipvs/ipset/testing/fake_test.go
@@ -20,7 +20,7 @@ import (
 	"testing"
 
 	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/kubernetes/pkg/util/ipset"
+	"k8s.io/kubernetes/pkg/proxy/ipvs/ipset"
 )
 
 const testVersion = "v6.19"
