--- pkg/util/ipset/OWNERS
@@ -1,11 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-reviewers:
-  - sig-network-reviewers
-approvers:
-  - sig-network-approvers
-labels:
-  - sig/network
-emeritus_approvers:
-  - brendandburns
-  - m1093782566
